Implemented a basic integration of the admin context module into the waddlebot DBM.
The admin context module creates a context session that contains an expiry date and a token. This is sent along with the marketplace modules when a command is queried.

On the listener side, if certain commands require admin privilages, a valid session token must be sent to the listener, else the command will fail