Implemented the alias module into the DBM.
This new module allows aliases to be added to commands so that certain commands can be executed using different commands set by a user.

Part of the integration is into the marketplace_modules, where if a given command is not found, the system attempts to find the command as an alias to send it back to the user.