Reworked all listener functions in each controller to follow a new payload structure, so that the given payload always contains the following values:
- identity_name
- community_name
- command_string
- channel_id
- account

All necessary command variables and parameters that are not the identity, community, channel or account values, are found within the command_string variable that is typed by the user. All the others are auto generated by the listener.

This follows a new structure that simplifies how commands are stored and processed by the listener, as well as the supporting the new module_commands table

This commit also includes several bugfixes